requirements.txt
apex==0.9.10dev
ConfigArgParse==1.7
cuml==0.6.1.post1
fvcore==0.1.5.post20221221
imageio==2.33.0
matplotlib==3.8.2
mmcv==2.1.0
mmcv_full==1.7.2
mmdet==3.3.0
mmsegmentation==1.2.2
numpy==1.24.1
omegaconf==2.3.0
opencv_python==4.8.1.78
opt_einsum==3.3.0
Pillow==10.1.0
Pillow==9.3.0
Pillow==10.3.0
pytorch_lightning==2.1.2
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.13.0
setuptools==60.2.0
skimage==0.0
submitit==1.5.1
timm==0.5.4
torch==2.1.0+cu118
torchmetrics==1.2.1
torchvision==0.16.0
tqdm==4.65.2
xformers==0.0.22.post7+cu118
